+++ Kventin.Backend/Kventin.Services/Dtos/Exercises/UpdateExerciseDto.cs
@@ -8,12 +8,12 @@
 
         public bool? IsIndividual { get; set; }
 
-        public long? TeacherId { get; set; }
+        public int? TeacherId { get; set; }
 
-        public long? StudyGroupId { get; set; }
+        public int? StudyGroupId { get; set; }
 
-        public long? LessonId { get; set; }
+        public int? LessonId { get; set; }
 
-        public long? IndividualStudentId { get; set; }
+        public int? IndividualStudentId { get; set; }
     }
 }
