+++ Kventin.Backend/Kventin.Services/Dtos/Files/FileInfoDto.cs
@@ -18,7 +18,7 @@ namespace Kventin.Services.Dtos.Files
         }
 
         [Required]
-        public long FileId { get; set; }
+        public int FileId { get; set; }
 
         [Required]
         public string ContentType { get; set; }
