+++ Kventin.Backend/Kventin.Services/Dtos/Marks/AssignMarksForExerciseDto.cs
@@ -5,7 +5,7 @@ namespace Kventin.Services.Dtos.Marks
     public class AssignMarksForExerciseDto
     {
         [Required]
-        public required long ExerciseId { get; set; }
+        public required int ExerciseId { get; set; }
 
         public List<StudentMarksDto> StudentMarks { get; set; } = [];
     }
