+++ Kventin.Backend/Kventin.Services/Dtos/Marks/AssignMarksForLessonDto.cs
@@ -5,7 +5,7 @@ namespace Kventin.Services.Dtos.Marks
     public class AssignMarksForLessonDto
     {
         [Required]
-        public long LessonId { get; set; }
+        public int LessonId { get; set; }
 
         public List<StudentMarksDto> StudentMarks { get; set; } = [];
     }
