+++ Kventin.Backend/Kventin.Services/Dtos/Schedule/AddScheduleItemDto.cs
@@ -16,16 +16,16 @@ namespace Kventin.Services.Dtos.Schedule
         public string? Classroom { get; set; }
 
         [Required]
-        public required long TeacherId { get; set; }
+        public required int TeacherId { get; set; }
 
         [Required]
-        public required long GroupId { get; set; }
+        public required int GroupId { get; set; }
 
         [Required]
-        public required long SubjectId { get; set; }
+        public required int SubjectId { get; set; }
 
         [Required]
-        public required long ScheduleId { get; set; }
+        public required int ScheduleId { get; set; }
 
         [Required]
         public required bool IsOnline { get; set; }
