+++ Kventin.Backend/Kventin.Services/Dtos/StudyGroups/CreateStudyGroupDto.cs
@@ -1,5 +1,4 @@
 ﻿using System.ComponentModel.DataAnnotations;
-using System.Runtime.CompilerServices;
 
 namespace Kventin.Services.Dtos.StudyGroups
 {
@@ -9,9 +8,9 @@ namespace Kventin.Services.Dtos.StudyGroups
         public required string GroupName { get; set; }
 
         [Required]
-        public required long SubjectId { get; set; }
+        public required int SubjectId { get; set; }
 
         [Required]
-        public required long TeacherId { get; set; }
+        public required int TeacherId { get; set; }
     }
 }
